requirements.txt
rocketapi
progress
colorama
python-telegram-bot
Pillow